Install dlt-viewer scripts using homedir
A previous method used a script to create some files on the fly
using parameterized location of Qt and dlt-viewer.

That was then removed, but that left a buggy setup.

For now, just put back the necessary files as they are under homedir and
insert the known absolute path (/home/vagrant/...)
Remove the cleverness and keep it simple - If $HOME changes later, rewrite
these absolute paths.